refactor: 모집 정보 저장 로직 분리
